Updated to_do_list.html
added more features like a  "completed" section for completed tasks and also a "✔️" button that will add a task to "completed" section with time. users won`t be able to delete the tasks that are newly added  to completed section. they need to refresh the page for the same thus prevents data loss if delete button is clicked mistakenly.